+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 更改日志 | CHANGE LOG
 
+- 更换背景图片功能支持设置无背景图片 | The background image replacement function supports setting no background image.
+
 ## v1.8.0/2023-04-12
 
 - [v1.7.6 <=> v1.8.0](https://github.com/Zuoqiu-Yingyi/siyuan-theme-dark-plus/compare/v1.7.6...v1.8.0)